# prototype-api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/labels": {
      "post": {
        "operationId": "createLabel",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLabelDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a Label.",
        "tags": [
          "labels"
        ]
      },
      "get": {
        "operationId": "listLabels",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LabelResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List Label records.",
        "tags": [
          "labels"
        ]
      }
    },
    "/labels/{id}": {
      "get": {
        "operationId": "getLabel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Label ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Label not found."
          }
        },
        "summary": "Get a Label by ID.",
        "tags": [
          "labels"
        ]
      },
      "patch": {
        "operationId": "updateLabel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Label ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLabelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Label not found."
          }
        },
        "summary": "Update a Label.",
        "tags": [
          "labels"
        ]
      },
      "delete": {
        "operationId": "deleteLabel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Label ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Label deleted."
          },
          "404": {
            "description": "Label not found."
          }
        },
        "summary": "Delete a Label.",
        "tags": [
          "labels"
        ]
      }
    },
...
    "/permissions": {
      "post": {
        "operationId": "createPermission",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a Permission.",
        "tags": [
          "permissions"
        ]
      },
      "get": {
        "operationId": "listPermissions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PermissionResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List Permission records.",
        "tags": [
          "permissions"
        ]
      }
    },
    "/permissions/{id}": {
      "get": {
        "operationId": "getPermission",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Permission ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Permission not found."
          }
        },
        "summary": "Get a Permission by ID.",
        "tags": [
          "permissions"
        ]
      },
      "patch": {
        "operationId": "updatePermission",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Permission ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Permission not found."
          }
        },
        "summary": "Update a Permission.",
        "tags": [
          "permissions"
        ]
      },
      "delete": {
        "operationId": "deletePermission",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Permission ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Permission deleted."
          },
          "404": {
            "description": "Permission not found."
          }
        },
        "summary": "Delete a Permission.",
        "tags": [
          "permissions"
        ]
      }
    },
    "/protocols": {
      "post": {
        "operationId": "createProtocol",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProtocolDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a Protocol.",
        "tags": [
          "protocols"
        ]
      },
      "get": {
        "operationId": "listProtocols",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProtocolResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List Protocol records.",
        "tags": [
          "protocols"
        ]
      }
    },
    "/protocols/{id}": {
      "get": {
        "operationId": "getProtocol",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Protocol ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Protocol not found."
          }
        },
        "summary": "Get a Protocol by ID.",
        "tags": [
          "protocols"
        ]
      },
      "patch": {
        "operationId": "updateProtocol",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Protocol ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProtocolDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Protocol not found."
          }
        },
        "summary": "Update a Protocol.",
        "tags": [
          "protocols"
        ]
      },
      "delete": {
        "operationId": "deleteProtocol",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Protocol ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Protocol deleted."
          },
          "404": {
            "description": "Protocol not found."
          }
        },
        "summary": "Delete a Protocol.",
        "tags": [
          "protocols"
        ]
      }
    },
...
  },
  "info": {
    "title": "IFS Prototype API",
    "description": "OpenAPI contract for the IFS prototype backend.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PermissionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Globally unique identifier for this permission."
          },
          "ifsId": {
            "type": "string",
            "description": "IFS system identifier for this Permission."
          },
          "entityType": {
            "type": "string",
            "description": "Entity type discriminator. Always \"Permission\" for Permission entities.",
            "enum": [
              "Permission"
            ]
          },
          "basedOn": {
            "type": "string",
            "description": "Id of the object this object is derived from."
          },
          "entityDocumentationUrl": {
            "type": "string",
            "description": "URL for documentation about this entity.",
            "format": "uri"
          },
          "actionIds": {
            "description": "ID list of actions that this permission allows",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "actions": {
            "description": "List of actions that this permission allows",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActionResponseDto"
            }
          },
          "memberId": {
            "type": "string",
            "description": "Identifier of the member receiving this permission."
          },
          "roleId": {
            "type": "string",
            "description": "Identifier of the role receiving this permission."
          },
          "scopeId": {
            "type": "string",
            "description": "Identifier of the scope object where this permission applies."
          },
          "scope": {
            "type": "object",
            "additionalProperties": true,
            "description": "Scope object describing the system, project, process, area, or other bounded context where this permission applies."
          },
          "state": {
            "enum": [
              "granted",
              "under-review",
              "revoked",
              "drafted"
            ],
            "type": "string",
            "description": "Lifecycle state of the permission. Example values may include drafted, under-review, active, suspended, or revoked."
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when this permission record was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when this permission record was last updated.",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "description": "Optional timestamp after which this permission no longer applies.",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "ifsId",
          "entityType",
          "entityDocumentationUrl",
          "actionIds",
          "scopeId",
          "state",
          "createdAt",
          "expiresAt"
        ]
      },
...
      "CreateLabelDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Globally unique identifier for this label."
          },
          "ifsId": {
            "type": "string",
            "description": "IFS system identifier for this Label."
          },
          "entityType": {
            "type": "string",
            "description": "Entity type discriminator. Always \"Label\" for Label entities.",
            "enum": [
              "Label"
            ]
          },
          "basedOn": {
            "type": "string",
            "description": "Id of the object this object is derived from."
          },
          "entityDocumentationUrl": {
            "type": "string",
            "description": "URL for documentation about this entity.",
            "format": "uri"
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when this label was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when this label was last updated.",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "ifsId",
          "entityType",
          "entityDocumentationUrl",
          "createdAt",
          "updatedAt"
        ]
      },
      "LabelResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Globally unique identifier for this label."
          },
          "ifsId": {
            "type": "string",
            "description": "IFS system identifier for this Label."
          },
          "entityType": {
            "type": "string",
            "description": "Entity type discriminator. Always \"Label\" for Label entities.",
            "enum": [
              "Label"
            ]
          },
          "basedOn": {
            "type": "string",
            "description": "Id of the object this object is derived from."
          },
          "entityDocumentationUrl": {
            "type": "string",
            "description": "URL for documentation about this entity.",
            "format": "uri"
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when this label was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when this label was last updated.",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "ifsId",
          "entityType",
          "entityDocumentationUrl",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateLabelDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Globally unique identifier for this label."
          },
          "ifsId": {
            "type": "string",
            "description": "IFS system identifier for this Label."
          },
          "entityType": {
            "type": "string",
            "description": "Entity type discriminator. Always \"Label\" for Label entities.",
            "enum": [
              "Label"
            ]
          },
          "basedOn": {
            "type": "string",
            "description": "Id of the object this object is derived from."
          },
          "entityDocumentationUrl": {
            "type": "string",
            "description": "URL for documentation about this entity.",
            "format": "uri"
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when this label was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when this label was last updated.",
            "format": "date-time"
          }
        }
      },
...
      "CreatePermissionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Globally unique identifier for this permission."
          },
          "ifsId": {
            "type": "string",
            "description": "IFS system identifier for this Permission."
          },
          "entityType": {
            "type": "string",
            "description": "Entity type discriminator. Always \"Permission\" for Permission entities.",
            "enum": [
              "Permission"
            ]
          },
          "basedOn": {
            "type": "string",
            "description": "Id of the object this object is derived from."
          },
          "entityDocumentationUrl": {
            "type": "string",
            "description": "URL for documentation about this entity.",
            "format": "uri"
          },
          "actionIds": {
            "description": "ID list of actions that this permission allows",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "actions": {
            "description": "List of actions that this permission allows",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActionResponseDto"
            }
          },
          "memberId": {
            "type": "string",
            "description": "Identifier of the member receiving this permission."
          },
          "roleId": {
            "type": "string",
            "description": "Identifier of the role receiving this permission."
          },
          "scopeId": {
            "type": "string",
            "description": "Identifier of the scope object where this permission applies."
          },
          "scope": {
            "type": "object",
            "additionalProperties": true,
            "description": "Scope object describing the system, project, process, area, or other bounded context where this permission applies."
          },
          "state": {
            "enum": [
              "granted",
              "under-review",
              "revoked",
              "drafted"
            ],
            "type": "string",
            "description": "Lifecycle state of the permission. Example values may include drafted, under-review, active, suspended, or revoked."
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when this permission record was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when this permission record was last updated.",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "description": "Optional timestamp after which this permission no longer applies.",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "ifsId",
          "entityType",
          "entityDocumentationUrl",
          "actionIds",
          "scopeId",
          "state",
          "createdAt",
          "expiresAt"
        ]
      },
      "UpdatePermissionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Globally unique identifier for this permission."
          },
          "ifsId": {
            "type": "string",
            "description": "IFS system identifier for this Permission."
          },
          "entityType": {
            "type": "string",
            "description": "Entity type discriminator. Always \"Permission\" for Permission entities.",
            "enum": [
              "Permission"
            ]
          },
          "basedOn": {
            "type": "string",
            "description": "Id of the object this object is derived from."
          },
          "entityDocumentationUrl": {
            "type": "string",
            "description": "URL for documentation about this entity.",
            "format": "uri"
          },
          "actionIds": {
            "description": "ID list of actions that this permission allows",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "actions": {
            "description": "List of actions that this permission allows",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActionResponseDto"
            }
          },
          "memberId": {
            "type": "string",
            "description": "Identifier of the member receiving this permission."
          },
          "roleId": {
            "type": "string",
            "description": "Identifier of the role receiving this permission."
          },
          "scopeId": {
            "type": "string",
            "description": "Identifier of the scope object where this permission applies."
          },
          "scope": {
            "type": "object",
            "additionalProperties": true,
            "description": "Scope object describing the system, project, process, area, or other bounded context where this permission applies."
          },
          "state": {
            "enum": [
              "granted",
              "under-review",
              "revoked",
              "drafted"
            ],
            "type": "string",
            "description": "Lifecycle state of the permission. Example values may include drafted, under-review, active, suspended, or revoked."
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when this permission record was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when this permission record was last updated.",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "description": "Optional timestamp after which this permission no longer applies.",
            "format": "date-time"
          }
        }
      },
      "CreateProtocolDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "identifier protocol."
          },
          "ifsId": {
            "type": "string",
            "description": "IFS system identifier for this Protocol."
          },
          "entityType": {
            "type": "string",
            "description": "Entity type discriminator. Always \"Protocol\" for Protocol entities.",
            "enum": [
              "Protocol"
            ]
          },
          "basedOn": {
            "type": "string",
            "description": "Id of the object this object is derived from."
          },
          "entityDocumentationUrl": {
            "type": "string",
            "description": "URL for documentation about this entity.",
            "format": "uri"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "ifsId",
          "entityType",
          "entityDocumentationUrl",
          "createdAt"
        ]
      },
      "ProtocolResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "identifier protocol."
          },
          "ifsId": {
            "type": "string",
            "description": "IFS system identifier for this Protocol."
          },
          "entityType": {
            "type": "string",
            "description": "Entity type discriminator. Always \"Protocol\" for Protocol entities.",
            "enum": [
              "Protocol"
            ]
          },
          "basedOn": {
            "type": "string",
            "description": "Id of the object this object is derived from."
          },
          "entityDocumentationUrl": {
            "type": "string",
            "description": "URL for documentation about this entity.",
            "format": "uri"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "ifsId",
          "entityType",
          "entityDocumentationUrl",
          "createdAt"
        ]
      },
...
    }
  }
}