# prototype-api/src/label/label.controller.ts
// Generated by scripts/generate-resource.mjs.
import {
  Body,
  Controller,
  Delete,
  Get,
  HttpCode,
  HttpStatus,
  NotFoundException,
  Param,
  Patch,
  Post,
} from '@nestjs/common';
import {
  ApiBody,
  ApiCreatedResponse,
  ApiNoContentResponse,
  ApiNotFoundResponse,
  ApiOkResponse,
  ApiOperation,
  ApiParam,
  ApiTags,
} from '@nestjs/swagger';
import { CreateLabelDto } from './dto/create-label.dto.js';
import { LabelResponseDto } from './dto/label-response.dto.js';
import { UpdateLabelDto } from './dto/update-label.dto.js';
import { LabelService } from './label.service.js';

@ApiTags('labels')
@Controller('labels')
export class LabelController {
  constructor(private readonly labelService: LabelService) {}

  @Post()
  @ApiOperation({ operationId: 'createLabel', summary: 'Create a Label.' })
  @ApiBody({ type: CreateLabelDto })
  @ApiCreatedResponse({ type: LabelResponseDto })
  create(@Body() data: CreateLabelDto): Promise<LabelResponseDto> {
    return this.labelService.create(data);
  }

  @Get()
  @ApiOperation({ operationId: 'listLabels', summary: 'List Label records.' })
  @ApiOkResponse({ type: LabelResponseDto, isArray: true })
  findAll(): Promise<LabelResponseDto[]> {
    return this.labelService.findAll();
  }

  @Get(':id')
  @ApiOperation({ operationId: 'getLabel', summary: 'Get a Label by ID.' })
  @ApiParam({ name: 'id', type: String, description: 'Label ID.' })
  @ApiOkResponse({ type: LabelResponseDto })
  @ApiNotFoundResponse({ description: 'Label not found.' })
  async findOne(@Param('id') id: string): Promise<LabelResponseDto> {
    const result = await this.labelService.findOne(id);
    if (!result) throw new NotFoundException('Label not found.');
    return result;
  }

  @Patch(':id')
  @ApiOperation({ operationId: 'updateLabel', summary: 'Update a Label.' })
  @ApiParam({ name: 'id', type: String, description: 'Label ID.' })
  @ApiBody({ type: UpdateLabelDto })
  @ApiOkResponse({ type: LabelResponseDto })
  @ApiNotFoundResponse({ description: 'Label not found.' })
  update(
    @Param('id') id: string,
    @Body() data: UpdateLabelDto,
  ): Promise<LabelResponseDto> {
    return this.labelService.update(id, data);
  }

  @Delete(':id')
  @HttpCode(HttpStatus.NO_CONTENT)
  @ApiOperation({ operationId: 'deleteLabel', summary: 'Delete a Label.' })
  @ApiParam({ name: 'id', type: String, description: 'Label ID.' })
  @ApiNoContentResponse({ description: 'Label deleted.' })
  @ApiNotFoundResponse({ description: 'Label not found.' })
  async remove(@Param('id') id: string): Promise<void> {
    await this.labelService.remove(id);
  }
}