# prototype-api/src/permission/permission.controller.ts
// Generated by scripts/generate-resource.mjs.
import {
  Body,
  Controller,
  Delete,
  Get,
  HttpCode,
  HttpStatus,
  NotFoundException,
  Param,
  Patch,
  Post,
} from '@nestjs/common';
import {
  ApiBody,
  ApiCreatedResponse,
  ApiNoContentResponse,
  ApiNotFoundResponse,
  ApiOkResponse,
  ApiOperation,
  ApiParam,
  ApiTags,
} from '@nestjs/swagger';
import { CreatePermissionDto } from './dto/create-permission.dto.js';
import { PermissionResponseDto } from './dto/permission-response.dto.js';
import { UpdatePermissionDto } from './dto/update-permission.dto.js';
import { PermissionService } from './permission.service.js';

@ApiTags('permissions')
@Controller('permissions')
export class PermissionController {
  constructor(private readonly permissionService: PermissionService) {}

  @Post()
  @ApiOperation({
    operationId: 'createPermission',
    summary: 'Create a Permission.',
  })
  @ApiBody({ type: CreatePermissionDto })
  @ApiCreatedResponse({ type: PermissionResponseDto })
  create(@Body() data: CreatePermissionDto): Promise<PermissionResponseDto> {
    return this.permissionService.create(data);
  }

  @Get()
  @ApiOperation({
    operationId: 'listPermissions',
    summary: 'List Permission records.',
  })
  @ApiOkResponse({ type: PermissionResponseDto, isArray: true })
  findAll(): Promise<PermissionResponseDto[]> {
    return this.permissionService.findAll();
  }

  @Get(':id')
  @ApiOperation({
    operationId: 'getPermission',
    summary: 'Get a Permission by ID.',
  })
  @ApiParam({ name: 'id', type: String, description: 'Permission ID.' })
  @ApiOkResponse({ type: PermissionResponseDto })
  @ApiNotFoundResponse({ description: 'Permission not found.' })
  async findOne(@Param('id') id: string): Promise<PermissionResponseDto> {
    const result = await this.permissionService.findOne(id);
    if (!result) throw new NotFoundException('Permission not found.');
    return result;
  }

  @Patch(':id')
  @ApiOperation({
    operationId: 'updatePermission',
    summary: 'Update a Permission.',
  })
  @ApiParam({ name: 'id', type: String, description: 'Permission ID.' })
  @ApiBody({ type: UpdatePermissionDto })
  @ApiOkResponse({ type: PermissionResponseDto })
  @ApiNotFoundResponse({ description: 'Permission not found.' })
  update(
    @Param('id') id: string,
    @Body() data: UpdatePermissionDto,
  ): Promise<PermissionResponseDto> {
    return this.permissionService.update(id, data);
  }

  @Delete(':id')
  @HttpCode(HttpStatus.NO_CONTENT)
  @ApiOperation({
    operationId: 'deletePermission',
    summary: 'Delete a Permission.',
  })
  @ApiParam({ name: 'id', type: String, description: 'Permission ID.' })
  @ApiNoContentResponse({ description: 'Permission deleted.' })
  @ApiNotFoundResponse({ description: 'Permission not found.' })
  async remove(@Param('id') id: string): Promise<void> {
    await this.permissionService.remove(id);
  }
}