# prototype-api/src/protocol/protocol.controller.ts
// Generated by scripts/generate-resource.mjs.
import {
  Body,
  Controller,
  Delete,
  Get,
  HttpCode,
  HttpStatus,
  NotFoundException,
  Param,
  Patch,
  Post,
} from '@nestjs/common';
import {
  ApiBody,
  ApiCreatedResponse,
  ApiNoContentResponse,
  ApiNotFoundResponse,
  ApiOkResponse,
  ApiOperation,
  ApiParam,
  ApiTags,
} from '@nestjs/swagger';
import { CreateProtocolDto } from './dto/create-protocol.dto.js';
import { ProtocolResponseDto } from './dto/protocol-response.dto.js';
import { UpdateProtocolDto } from './dto/update-protocol.dto.js';
import { ProtocolService } from './protocol.service.js';

@ApiTags('protocols')
@Controller('protocols')
export class ProtocolController {
  constructor(private readonly protocolService: ProtocolService) {}

  @Post()
  @ApiOperation({
    operationId: 'createProtocol',
    summary: 'Create a Protocol.',
  })
  @ApiBody({ type: CreateProtocolDto })
  @ApiCreatedResponse({ type: ProtocolResponseDto })
  create(@Body() data: CreateProtocolDto): Promise<ProtocolResponseDto> {
    return this.protocolService.create(data);
  }

  @Get()
  @ApiOperation({
    operationId: 'listProtocols',
    summary: 'List Protocol records.',
  })
  @ApiOkResponse({ type: ProtocolResponseDto, isArray: true })
  findAll(): Promise<ProtocolResponseDto[]> {
    return this.protocolService.findAll();
  }

  @Get(':id')
  @ApiOperation({
    operationId: 'getProtocol',
    summary: 'Get a Protocol by ID.',
  })
  @ApiParam({ name: 'id', type: String, description: 'Protocol ID.' })
  @ApiOkResponse({ type: ProtocolResponseDto })
  @ApiNotFoundResponse({ description: 'Protocol not found.' })
  async findOne(@Param('id') id: string): Promise<ProtocolResponseDto> {
    const result = await this.protocolService.findOne(id);
    if (!result) throw new NotFoundException('Protocol not found.');
    return result;
  }

  @Patch(':id')
  @ApiOperation({
    operationId: 'updateProtocol',
    summary: 'Update a Protocol.',
  })
  @ApiParam({ name: 'id', type: String, description: 'Protocol ID.' })
  @ApiBody({ type: UpdateProtocolDto })
  @ApiOkResponse({ type: ProtocolResponseDto })
  @ApiNotFoundResponse({ description: 'Protocol not found.' })
  update(
    @Param('id') id: string,
    @Body() data: UpdateProtocolDto,
  ): Promise<ProtocolResponseDto> {
    return this.protocolService.update(id, data);
  }

  @Delete(':id')
  @HttpCode(HttpStatus.NO_CONTENT)
  @ApiOperation({
    operationId: 'deleteProtocol',
    summary: 'Delete a Protocol.',
  })
  @ApiParam({ name: 'id', type: String, description: 'Protocol ID.' })
  @ApiNoContentResponse({ description: 'Protocol deleted.' })
  @ApiNotFoundResponse({ description: 'Protocol not found.' })
  async remove(@Param('id') id: string): Promise<void> {
    await this.protocolService.remove(id);
  }
}